requirements.txt
# Core Dependencies
numpy>=1.24.0
scipy>=1.10.0
sympy>=1.12

# Image Processing and Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0
imageio>=2.31.0
tifffile>=2023.7.0

# Web Scraping and Data Collection
scrapy>=2.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.12.0

# IIIF Support
iiif>=1.0.0
piffle>=0.4.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# API and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Data Analysis and Machine Learning
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Pattern Recognition and ML
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# Topological Data Analysis
gudhi>=3.8.0

# Sound and Acoustic Analysis
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Geometric Analysis
shapely>=2.0.0
networkx>=3.1

# Image Hashing and Deduplication
imagehash>=4.3.0

# CLI and UI
click>=8.1.0
rich>=13.5.0
typer>=0.9.0
streamlit>=1.28.0
plotly>=5.17.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Parallel Processing
joblib>=1.3.0
dask>=2023.9.0

# File Format Support
h5py>=3.9.0
pyarrow>=13.0.0

# Optical Character Recognition
pytesseract>=0.3.10
pdf2image>=1.16.0

# API Keys Management
keyring>=24.2.0
